--- .flaskenv
@@ -1,3 +1,0 @@
-FLASK_APP=run.py
-FLASK_ENV=development
-SECRET_KEY='sanket'
