--- run.py
@@ -1,10 +1,0 @@
-from app import create_app
-from config import load_config
-
-"""To start flask application run this file."""
-
-if __name__ == "__main__":
-    config = load_config()
-    app = create_app(config)
-    PORT = 5000
-    app.run(port=PORT)
