+++ server/nginx.conf
@@ -1,0 +1,17 @@
+upstream backend{
+    server web:5000;
+}
+
+server {
+    listen 80;
+
+    access_log  /var/log/nginx/access.log;
+    error_log  /var/log/nginx/error.log;
+
+    location / {
+        proxy_pass  http://backend;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header Host $host;
+        proxy_redirect off;
+    }
+}
