+++ web/.env
@@ -1,0 +1,5 @@
+FLASK_APP=run.py
+SECRET_KEY=sanket
+MONGODB_DB=BooksDB
+MONGODB_HOST=localhost
+MONGODB_PORT=27016
